Check if figure dir exists before launching Ipe
Ipe will launch fine with a file in a dir that does not exist, but will error when trying to save.

This checks the dir exists and creates it if not.

This also fixes a bug where the name for the figureDirectory setting hadn't been updated and so the setting was useless.